Refactor : 필터 기능 코드 정리 및 수정
